+++ src/InFlightPacket.hpp
@@ -17,19 +17,19 @@ class DeliveryNotificationManager;
 class InFlightPacket
 {
 public:
-    InFlightPacket(uint16_t sequenceNumber, float timeDispatched);
+    InFlightPacket(uint16_t sequenceNumber, uint32_t timeDispatched);
     ~InFlightPacket();
     
     void setTransmissionData(int key, TransmissionData* transmissionData);
     TransmissionData* getTransmissionData(int key) const;
     uint16_t getSequenceNumber() const;
-    float getTimeDispatched() const;
+    uint32_t getTimeDispatched() const;
     void handleDeliveryFailure(DeliveryNotificationManager* dnm) const;
     void handleDeliverySuccess(DeliveryNotificationManager* dnm) const;
     
 private:
     uint16_t _sequenceNumber;
-    float _timeDispatched;
+    uint32_t _timeDispatched;
     int _key;
     
     std::map<int, TransmissionData*> _transmissionDataMap;
